Add `_later_bulk` methods on non-Active Record objects.
Having a `_later_bulk` methods on non-Active Records can still be pretty nifty, so we should allow that.

Then if we're on Active Records we can default to `all` records if no argument is passed, but we can support passing in a subset directly too.

Then on non-Active Records, we can require the argument.